arrow beta
